Realizovana obrada rezervacija bez notifikaije. Implementirana agregatna funkcija brojanja rezervacija za jedan restoran za konkretan mjesec

+++ src/main/java/com/example/backend4rate/services/impl/ReservationService.java
@@ -1,24 +1,18 @@
 package com.example.backend4rate.services.impl;
 
-import java.time.LocalDateTime;
+import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
 
 import org.modelmapper.ModelMapper;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
 import com.example.backend4rate.exceptions.DuplicateReservationException;
 import com.example.backend4rate.exceptions.NotFoundException;
 import com.example.backend4rate.models.dto.Reservation;
-import com.example.backend4rate.models.dto.UpdateRestaurant;
-import com.example.backend4rate.models.dto.UserAccountResponse;
-import com.example.backend4rate.models.entities.CategoryEntity;
 import com.example.backend4rate.models.entities.GuestEntity;
 import com.example.backend4rate.models.entities.ReservationEntity;
-import com.example.backend4rate.models.entities.RestaurantCategoryEntity;
 import com.example.backend4rate.models.entities.RestaurantEntity;
-import com.example.backend4rate.models.entities.RestaurantPhoneEntity;
 import com.example.backend4rate.repositories.CategoryRepository;
 import com.example.backend4rate.repositories.GuestRepository;
 import com.example.backend4rate.repositories.ReservationRepository;
@@ -27,7 +21,6 @@ import com.example.backend4rate.repositories.RestaurantPhoneRepository;
 import com.example.backend4rate.repositories.RestaurantRepository;
 import com.example.backend4rate.services.ReservationServiceInterface;
 
-import jakarta.transaction.Transactional;
 
 @Service
 public class ReservationService implements ReservationServiceInterface {
@@ -37,9 +30,6 @@ public class ReservationService implements ReservationServiceInterface {
 
     private final GuestRepository guestRepository;
     private final RestaurantRepository restaurantRepository;
-    private final RestaurantPhoneRepository restaurantPhoneRepository;
-    private final RestaurantCategoryRepository restaurantCategoryRepository;
-    private final CategoryRepository categoryRepository;
 
     public ReservationService(ReservationRepository reservationRepository, GuestRepository guestRepository,
             RestaurantRepository restaurantRepository, ModelMapper modelMapper,
@@ -49,9 +39,6 @@ public class ReservationService implements ReservationServiceInterface {
         this.modelMapper = modelMapper;
         this.guestRepository = guestRepository;
         this.restaurantRepository = restaurantRepository;
-        this.restaurantPhoneRepository = restaurantPhoneRepository;
-        this.categoryRepository = categoryRepository;
-        this.restaurantCategoryRepository = restaurantCategoryRepository;
     }
 
     @Override
@@ -60,10 +47,12 @@ public class ReservationService implements ReservationServiceInterface {
                 Reservation.class);
     }
 
-    @Override
-    public void deleteReservation(Integer reservationId) {
-        reservationRepository.deleteById(reservationId);
-    }
+   /* @Override
+    public void deleteReservation(Integer reservationId) throws NotFoundException {
+        ReservationEntity reservationEntity= reservationRepository.findById(reservationId).orElseThrow(NotFoundException::new);
+        if ("canceled".equals(reservationEntity.getStatus()) || "denied".equals(reservationEntity.getStatus()))
+            reservationRepository.deleteById(reservationId);
+    }*/
 
     @Override
     public Reservation makeReservation(Reservation reservation)
@@ -77,6 +66,7 @@ public class ReservationService implements ReservationServiceInterface {
             throw new DuplicateReservationException("Reservation on this date is already made!");
         ReservationEntity reservationEntity = modelMapper.map(reservation, ReservationEntity.class);
         reservationEntity.setId(null);
+        reservationEntity.setStatus("pending");
         reservationEntity = reservationRepository.saveAndFlush(reservationEntity);
 
         return modelMapper.map(reservationEntity, Reservation.class);
@@ -87,7 +77,7 @@ public class ReservationService implements ReservationServiceInterface {
         List<ReservationEntity> reservationEntityList = reservationRepository.findAllByGuest_Id(guestId);
         if (reservationEntityList.isEmpty())
             throw new NotFoundException("Guest hasn't made any reservations! ");
-        return reservationEntityList.stream().map(l -> modelMapper.map(l, Reservation.class))
+        return reservationEntityList.stream().filter(l -> (l.getStatus().equals("pending") || l.getStatus().equals("approved")) && l.getDate().after(new Date())).map(l -> modelMapper.map(l, Reservation.class))
                 .collect(Collectors.toList());
     }
 
@@ -96,15 +86,47 @@ public class ReservationService implements ReservationServiceInterface {
         List<ReservationEntity> reservationEntityList = reservationRepository.findAllByRestaurant_Id(restaurantId);
         if (reservationEntityList.isEmpty())
             throw new NotFoundException("There are no reservations for this restaurant!");
-        return reservationEntityList.stream().map(l -> modelMapper.map(l, Reservation.class))
+        return reservationEntityList.stream().filter(l -> !l.getStatus().equals("denied") && l.getDate().after(new Date())).map(l -> modelMapper.map(l, Reservation.class))
                 .collect(Collectors.toList());
     }
 
-    @Scheduled(fixedRate = 3600000)
+    /*@Scheduled(fixedRate = 3600000)
     @Transactional
     public void deleteExpiredReservations() {
         LocalDateTime now = LocalDateTime.now();
         reservationRepository.deleteByDateBefore(now);
+    }*/
+
+    @Override
+    public Reservation approveReservation(Integer reservationId) throws NotFoundException {
+       ReservationEntity reservationEntity = reservationRepository.findById(reservationId).orElseThrow(NotFoundException::new);
+       reservationEntity.setStatus("approved");
+       reservationRepository.saveAndFlush(reservationEntity);
+       return modelMapper.map(reservationEntity, Reservation.class);
+    }
+
+    @Override
+    public Reservation denyReservation(Integer reservationId) throws NotFoundException {
+        ReservationEntity reservationEntity = reservationRepository.findById(reservationId).orElseThrow(NotFoundException::new);
+        reservationEntity.setStatus("denied");
+        reservationRepository.saveAndFlush(reservationEntity);
+        //TO-DO Obavijesti gosta o rezultatu obrade
+        return modelMapper.map(reservationEntity, Reservation.class);
+    }
+
+    @Override
+    public Reservation cancelReservation(Integer reservationId) throws NotFoundException {
+        ReservationEntity reservationEntity = reservationRepository.findById(reservationId).orElseThrow(NotFoundException::new);
+        reservationEntity.setStatus("canceled");
+        reservationRepository.saveAndFlush(reservationEntity);
+        //TO-DO Obavijesti gosta o rezultatu obrade
+        return modelMapper.map(reservationEntity, Reservation.class);
+
+    }
+
+    @Override
+    public Long numberOfReservationsByMonth(Integer restaurantId, Integer month, Integer year) {
+        return reservationRepository.countReservationsByRestaurantAndMonthAndYear(restaurantId, month, year);
     }
 
 }
